// src/app/_restClientes/model/cliente.ts
/**
 * Servicio Clientes Kallsonys
 * Servicio para la gestión de clientes en  OMS Kallsonys
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Clientes de la plataforma Kallsonys
 */
export interface Cliente { 
    idCliente?: number;
    usuario?: string;
    password?: string;
    apellido?: string;
    nombre?: string;
    telefono?: string;
    email?: string;
    estado?: string;
    idCategoria?: number;
    origen?: string;
    tipoidentificacion?: string;
    numidentificacion?: string;
}