// src/app/_restMail/model/correos.ts
/**
 * Servicio MailSend Kallsonys
 * Servicio para el envio de correos
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface Correos { 
    destinatario?: string;
    cuerpo?: string;
    mensaje?: string;
}